Refactor methods to position them in correct dirs
