+++ composables/services/admin/category/categoryArrangeService.ts
@@ -1,5 +1,4 @@
 import { where } from 'firebase/firestore'
-import { useYimaAdminCategory } from '~/composables/services/admin/category'
 import { queryByCollection } from '~/server/lib/firestore'
 import { ClientCategoryCache } from '~/server/lib/clientCategoryCache'
 
@@ -14,7 +13,7 @@ const CATEGORY_COLLECTION = 'category'
 export async function storeRepositioned(categories: AdminCategory[]) {
   fillPositionPropertyByOrder(categories)
   const updatePromises = categories.map(async (category) => {
-    return setCategory(category.id, { position: category.position })
+    await setCategory(category.id, { position: category.position! })
   })
 
   await Promise.all(updatePromises)
@@ -48,29 +47,6 @@ export function fillPositionPropertyByOrder(categories: AdminCategory[]) {
   let index = 1
   for (const category of categories) {
     category.position = index++
-  }
-
-  return categories
-}
-
-export function sortRelativesByPosition<C extends Category>(categories: C[]): C[] {
-  return categories[0]?.position === undefined
-    ? categories.sort((cat1, cat2) => cat1.name.localeCompare(cat2.name))
-    : categories.sort((cat1, cat2) => cat1.position! - cat2.position!)
-}
-
-export function deepSortRelativesByPosition<C extends Category>(categories: C[]): C[] {
-  categories = sortRelativesByPosition(categories)
-  for (const category of categories) {
-    if (category.children === undefined) {
-      continue
-    }
-
-    if (category.children.length === 0) {
-      continue
-    }
-
-    category.children = deepSortRelativesByPosition(category.children)
   }
 
   return categories
